+++ CHANGELOG.md
@@ -13,6 +13,9 @@
     -   Set a default `--push_delay` (#34)
 -   `auto-pr init`:
     -   Ask for the API key if not passed via parameters (#34)
+-   `auto-pr pull`:
+    -   Pull repositories in parallel (#36)
+    -   Add `-j` flag (#36)
 -   Add `auto-pr progress` command (#35)
 
 ## 0.2.0